Add files via upload
(R)ecent advances in deep brain stimulators and brain -machine interfaces have greatly expanded the possibilities of neuroprosthetics and neuromodulation. Together with advances in neuroengineering, nanotechnology, molecular biology and material sciences, it is now possible to address fundamental questions in neuroscience in new, more powerful ways. It is now possible to apply these new technologies in ways that range from augmenting and restoring function to neuromodulation modalities that treat neuropsychiatric disorders. Recent developments in neuromodulation methods offer significant advantages and potential clinical benefits for a variety of disorders. Here we describe the current state of the art in neuromodulation methods, and some advances in brain -machine interfaces, describing the advantages and limitations of the clinical applications of each method. The future applications of these new methods and how they will shape the future of psychiatry and medicine, along with safety and ethical implications, are also discussed. (c) 2017 Elsevier Inc. All rights reserved.