Add files via upload
A two-dimensional Boussinesq model of thermal convection is investigated by numerical simulations. The turbulent heat transport is mainly due to thermal plumes. It is shown that their thin, wrinkled interfaces control the statistics of large temperature excursions. Away from the interfaces, the temperature fluctuations are strongly suppressed and the field in those regions is well mixed. Their statistical signature is in the low-order moments of the temperature increments, whose scaling exponents depend linearly on the order.