Add files via upload
Visualization and presentation techniques are experiencing a rapid development and application in the architectural profession and beyond. As such, Stereoscopic Images (SI) and Virtual Reality (VR), both advanced visualization techniques, have found their way in daylighting research. This paper explores the correspondence between these two techniques using both quantitative and qualitative methods. The study focused on two small rooms with the same dimensions, with alternately white and black surfaces, and three different windows sizes. Seven architectural qualities were studied: Pleasantness, Calmness, Interest, Excitement, Complexity, Spaciousness, and Amount of View. The attributes were evaluated by 20 participants using a Likert-type scale. The collected data was analyzed using Wilcoxon signed-rank tests and the Bland-Altman method. The results revealed that the attributes Pleasantness, Calmness, Interest, and Complexity in rooms with both white and black surfaces, and Excitement and Spaciousness in a room with white surfaces are evaluated similarly in both VR and SI. The attributes Excitement and Spaciousness in a room with black surfaces and the Amount of View in rooms with both white and black surfaces were not evaluated similarly with both methods. In addition, qualitative results indicated that the visualization technique affects how a space is perceived. Indeed, the Virtual Reality's nature as an immersive environment provides the feeling of ‘presence’ which does not apply to SI. The results of this paper can help researchers working with daylighting in buildings in selecting the appropriate visualization technique to reduce experimental and logistical constraints caused by varying daylight conditions.