Add files via upload
Zymoseptoria tritici, the causal agent of septoria tritici blotch, a serious foliar disease of wheat, is a necrotrophic pathogen that undergoes a long latent period. Emergence of insensitivity to fungicides, and pesticide reduction policies, mean there is a pressing need to understand septoria and control it through greater varietal resistance. Stb6 and Stb15, the most common qualitative resistance genes in modern wheat cultivars, determine specific resistance to avirulent fungal genotypes following a gene-for-gene relationship. This study investigated compatible and incompatible interactions of wheat with Z. tritici using eight combinations of cultivars and isolates, with the aim of identifying molecular responses that could be used as markers for disease resistance during the early, symptomless phase of colonization. The accumulation of TaMPK3 was estimated using western blotting, and the expression of genes implicated in gene-for-gene interactions of plants with a wide range of other pathogens was measured by qRT-PCR during the presymptomatic stages of infection. Production of TaMPK3 and expression of most of the genes responded to inoculation with Z. tritici but varied considerably between experimental replicates. However, there was no significant difference between compatible and incompatible interactions in any of the responses tested. These results demonstrate that the molecular biology of the gene-for-gene interaction between wheat and Zymoseptoria is unlike that in many other plant diseases, indicate that environmental conditions may strongly influence early responses of wheat to infection by Z. tritici, and emphasize the importance of including both compatible and incompatible interactions when investigating the biology of this complex pathosystem.