Add files via upload
Zwitterionic peptides were anchored to a conducting polymer of citrate doped poly(3,4-ethylenedioxythiophene) (PEDOT) via the nickel cation coordination, and the obtained peptide modified PEDOT, with excellent antifouling ability and good conductivity, was further used for the immobilization of a DNA probe to construct an electrochemical biosensor for the breast cancer marker BRCA1. The DNA biosensor was highly sensitive (with detection limit of 0.03 fM) and selective, and it was able to detect BRCA1 in 5% (v/v) human plasma with satisfying accuracy and low fouling. The marriage of antifouling and biocompatible peptides with conducting polymers opened a new avenue to construct electrochemical biosensors capable of assaying targets in complex biological media with high sensitivity and without biofouling.