Add files via upload
Zinc plays a very important role in animal and human metabolism. Nowadays, it is one of the most extensively studied trace element, since its sphere of action has been demonstrated to be very broad. From the biochemical standpoint, it controls more than 300 different enzymes, many of them involved with intermediary metabolism, DNA and RNA synthesis, gene expression, and immunocompetence. It also plays a significant role in hormonal homeostasis, since it can interact with almost all hormones. Zn2+ is closely related to the thyroid and steroid hormones, insulin, parathormone, and pituitary hormones, particularly prolactin (PRL). Zn2+ can inhibit PRL secretion within a range of physiologically and pharmacologically relevant concentrations. This property has raised the possibility of clinical applications of zinc. In this article, we review the literature on the subject in an attempt to provide a comprehensible general view.