Add files via upload
Objective: The purpose of this study was to evaluate serum zinc status of pregnant women in the China Adult Chronic Disease and Nutrition Surveillance (CACDNS) in 2015–2016. Methods: A total of 7147 apparently healthy pregnant women were randomly selected in 302 national monitoring sites. Information on age, race, residence region, education, pregnancy, and family income per annum was collected, and the concentration of serum zinc was determined. The evaluation of serum zinc status was further performed according to the recommendations by the International Zinc Nutrition Consultative Group (IZiNCG). Results: The median concentration of serum zinc was 858.9 μg/L with an interquartile range (IQR) of 712.9 μg/L and 1048.9 μg/L, while the overall prevalence of zinc deficiency was 3.5% with a 95% confidence interval (CI) of 3.0% and 3.9%. Serum zinc status of pregnant women changed greatly in the different categories, particular in pregnancy and family income per annum (p < 0.05), but no significant difference was observed in the prevalence of zinc deficiency (p > 0.05). Conclusions: The lower prevalence of zinc deficiency generally indicated a better zinc status for pregnant women in the CACDNS in 2015–2016. However, a well-designed evaluation system of zinc status for pregnant women should be continually optimized and improved by inducing more parameters such as biochemical, dietary, or functional indicators.