Add files via upload
Proper intake of dietary nutrients is considered crucial for preventing the initiation of events leading to the development of carcinoma. Many dietary compounds have been considered to contribute in cancer prevention including zinc, which plays a pivotal role in host defense against the initiation and promotion of several malignancies. Zinc is an essential element that is integral to many proteins and transcription factors which regulate key cellular functions such as the response to oxidative stress, DNA replication, DNA damage repair, cell cycle progression, and apoptosis. Zinc has been ascribed roles in the metabolism and interaction of malignant cells, particularly in apoptosis. Zinc is involved in structural stabilization and activation of the p53 that appears to be an important component of the apoptotic process and also in activation of certain members of the caspase family of proteases. Zinc exerts a positive beneficial effect against chemically induced preneoplastic progression in rats and provides an effective dietary chemopreventive approach to disease in vulnerable section of population with family history of carcinoma. The present review provides an insight into the research conducted on animals as well as on human subjects for providing the concept that zinc deficiency is an important factor in the development and progression of malignancy and that zinc could be efficacious in the prevention and treatment of several cancers viz., colon, pancreas, oesophageal and head and neck. However, it needs further exploration with regard to other definitive bioassays including protein expression and documentation of specific molecular markers to establish the exact mechanism for zinc-mediated cancer chemoprevention. Preclinical trials need to investigate the genetic and epigenetic pathways of chemoprevention by zinc.