Add files via upload
Zika virus (ZIKV) is an emerging mosquito-borne (Aedes genus) arbovirus of the Flaviviridae family. Following epidemics in Micronesia and French Polynesia during the past decade, more recent ZIKV infection outbreaks were first reported in South America as early as May of 2013, and spread to now 50 countries throughout the Americas. Although no other flavivirus has previously been known to cause major fetal malformations following perinatal infection, reports of a causal link between ZIKV and microcephaly, brain and ocular malformations, and fetal loss emerged from hard hit regions of Brazil by October 2015. Among the minority of infected women with symptoms, clinical manifestations of ZIKV infection may include fever, headache, arthralgia, myalgia and maculopapular rash; however, only one out of every four to five people who are infected have any symptoms. Thus, clinical symptom reporting is an ineffective screening tool for the relative risk assessment of ZIKV infection in the majority of patients. As previously occurred with other largely asymptomatic viral infections posing perinatal transmission risk (such as HIV or CMV), we must develop and implement rapid, sensitive, and specific screening and diagnostic testing for both viral detection and estimation of timing of exposure. Unfortunately, despite an unprecedented surge in attempts to rapidly advance perinatal clinical testing for a previously obscure arbovirus, there are several ongoing hindrances to molecular and sonographic based screening and diagnosis of congenital ZIKV infection. These include: (1) difficulty in estimating the timing of exposure for women living in endemic areas, and thus limited interpretability of IgM serologies; (2) crossreaction of IgM serologies with other endemic flaviruses, such as dengue (DENV); (3) persistent viremia and viruria in pregnancy weeks to months after primary exposure; and (4) fetal brain malformations and anomalies preceding the sonographic detection of microcephaly. In this commentary, we discuss screening and diagnostic considerations which are grounded not only in the realities of current obstetrical practice in a largely global population, but in basic immunology and virology. We review recent epidemiologic data pertaining to risk of congenital ZIKV malformations based on trimester of exposure, and consider side by side with emerging data demonstrating replication of ZIKV in placental and fetal tissue throughout gestation. We discuss limitations to ultrasound based strategies which rely largely or solely on the detection of microcephaly, and provide alternative neurosonographic approaches for the detection of malformations which may precede or occur independent of a small head circumference. This expert review provides information that is of value for the: 1) obstetrician, maternal-fetal medicine specialist, midwife, patient and family in cases of suspected ZIKV infection; 2) reviews the methodology for laboratory testing to explore the presence of the virus and the immune response; 3) ultrasound based assessment of the fetus suspected to be exposed to ZIKV with particular emphasis on the central nervous system; and 4) identifies areas ready for development.