Add files via upload
The Zika virus (ZIKV) has recently attracted major research interest as infection was unexpectedly associated with neurological manifestations in developing foetuses and with Guillain-Barré syndrome in infected adults. Understanding the underlying molecular mechanisms requires reverse genetic systems, which allow manipulation of infectious cDNA clones at will. In the case of flaviviruses, to which ZIKV belongs, several reports have indicated that the construction of full-length cDNA clones is difficult due to toxicity during plasmid amplification in Escherichia coli. Toxicity of flaviviral cDNAs has been linked to the activity of cryptic prokaryotic promoters within the region encoding the structural proteins leading to spurious transcription and expression of toxic viral proteins. Here, we employ an approach based on in silico prediction and mutational silencing of putative promoters to generate full-length cDNA clones of the historical MR766 strain and the contemporary French Polynesian strain H/PF/2013 of ZIKV. While for both strains construction of full-length cDNA clones has failed in the past, we show that our approach generates cDNA clones that are stable on single bacterial plasmids and give rise to infectious viruses with properties similar to those generated by other more complex assembly strategies. Further, we generate luciferase and fluorescent reporter viruses as well as sub-genomic replicons that are fully functional and suitable for various research and drug screening applications. Taken together, this study confirms that in silico prediction and silencing of cryptic prokaryotic promoters is an efficient strategy to generate full-length cDNA clones of flaviviruses and reports novel tools that will facilitate research on ZIKV biology and development of antiviral strategies.