Add files via upload
Zero-wear is considered as wear within the limits of surface topography of the component. The honed cylinder liner surface is composed of the plateaus and relatively deep valleys. In this paper, a two-scale homogenized mixed lubrication and wear model is established to study the "zero-wear" process of the cylinder liner. In order to consider the plateau roughness and the valleys separately on local and global scales, the honed surface is firstly decomposed into plateaus and valleys by the morphological filtering approach. Then, based on two-scale homogenization technique, a homogenized mixed model is developed to characterize the lubrication properties and the contact severity of the piston ring-liner system. On this basis, the cylinder liner "zero-wear" process is divided into two stages: at the first stage, running in of the plateau component on the local scale; at the second stage, wear of the valley component on the global scale. On the local scale, the influence of the plateau component running-in on the homogenized flow factors and contact stiffness is studied. On the global scale, the influence of the valley component wear on the lubrication properties (load carrying capacity, contact force and friction coefficient) is stu, died. Experiments are conducted on a reciprocating tester. The experimentally measured friction coefficient and wear depth are shown to be in good agreement with the simulation results. The results showed that the two-scale homogenized mixed lubrication and wear model developed in this paper can well predidt the cylinder liner "zero wear" process. (C) 2016 Elsevier B.V. All rights reserved.