Add files via upload
Zero-difference balanced (ZDB) functions have many applications in coding theory, cryptography and communications engineering and so on. Recently, the authors [10, 11] generalized the definition of ZDB functions to be G-ZDB functions. In this paper, based on p-cyclotomic cosets, two classes of ZDB functions are obtained, and several new classes of G-ZDB functions are constructed. Furthermore, some frequency-hopping sequences are obtained directly from G-ZDB functions.