Add files via upload
Zernike moments (ZMs) are very useful image descriptors which belong to a family of orthogonal rotation invariant moments. Due to their many attractive characteristics, they have been used in many pattern recognition, image processing and computer vision applications. However, they suffer from very high computation complexity which prohibits their use in many practical problems. The ZMs are computed as a convolution process between the image data and the Zernike kernel functions. In the past, various attempts have been made for the efficient computation of ZMs and considerable success has been achieved using recursive relations and 8-way symmetry/anti-symmetry of Zernike function. In this paper, we propose a new computational flow model for the convolution of the image data with the Zernike kernel functions. The proposed model also takes advantage of the 8-way symmetry/anti-symmetry property of the kernel function and builds up the convolution process which reduces the number of additions/subtractions from 56 to 24 and the number of multiplications from 12 to 8 (refer Table 6 in the text) for each location in an octant of a circular disk on which the moments are computed. Detailed experimental results show that the speed of the ZMs computation increases by a factor varying from 15% to 41% (depending upon the order of moments) for multiple images as compared to the existing fast algorithms available in the literature. When ZMs are computed at each pixel of an image on overlapping blocks, the improvement in computation time varies from 10% to 33%. (C) 2016 Elsevier GmbH. All rights reserved.